requirements.txt
Django
djangorestframework
PyJWT
mysqlclient
redis
django-cors-headers
pyotp
qrcode
dotenv
djangorestframework-simplejwt
Pillow